Major Style update
Changed some spacing throughout the document. Added a background image. Colorized and stylized the output. Added a button for copying the output into the clipboard. Added a way temporarily change the code used from the default. *note this change only lasts until the page is reloaded and then the default returns.